requirements.txt
python-telegram-bot==10.0.1
python-decouple==3.1